regresion/requirements.txt
selenium
fpdf2
pytest
